Logic was added to handle bottom nav bar and projects button.
