+++ 2. Application/src/components/WelcomeComp/WelcomeCompLogic.js
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from "react";
+import { useEffect, useRef } from 'react';
 
 import anime from 'animejs/lib/anime.es.js';
